Ajustando os namespaces de todo o Projeto

+++ SMF/Domain/Entities/Loteamento.cs
@@ -4,7 +4,7 @@
     {
         public int LoteamentoId { get; set; }
         public string? Descricao { get; set; }
-        public int? RegiaoId { get; set; }
+        public long? RegiaoId { get; set; }
         public Regiao? Regiao { get; set; }
         public DateTime? DataCadastro { get; set; }
